+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* User can set activeColor of checkbox for both parent and children
+* Added feature for having multiple ParentChildCheckbox
+* Can get result of checkboxes using map `isParentSelected` and `selectedChildrenMap`
+* Added null-safety
+